Disable RStudio authentication
**IMPORTANT**: This will start the server without authentication. Always
run this image behind an Auth Proxy (we do).

Why?
----
Currently we have to do [weird things] in a custom [rstudio-auth-proxy] in order to authenticate the user through RStudio Server.

The reason for disabling authentication is because this way we can
replace the custom [rstudio-auth-proxy] with the our generic [auth-proxy].
This will remove the need to manage two Auth Proxies which are doing
essentially the same thing.

Ticket
------

https://trello.com/c/5twZHrb4/215-try-to-replace-custom-rstudio-auth-proxy-with-our-usual-generic-one

[weird things]: https://github.com/ministryofjustice/analytics-platform-rstudio-auth-proxy/blob/8d4afb4982fddf3ad0d4446cca1e9f8455e160fe/app/auth.js
[rstudio-auth-proxy]: https://github.com/ministryofjustice/analytics-platform-rstudio-auth-proxy
[auth-proxy]: https://github.com/ministryofjustice/analytics-platform-auth-proxy